consolidates table header to be one external comp
