+++ components/common/Tooltip.js
@@ -2,7 +2,7 @@ import React from 'react'
 import InfoIcon from '../../static/circle-information'
 import ReactTooltip from 'react-tooltip'
 
-export const Tooltip = props => {
+const Tooltip = props => {
   return (
     <>
       <span data-tip={props.dataTip}>
@@ -13,3 +13,5 @@ export const Tooltip = props => {
     </>
   )
 }
+
+export default Tooltip
