+++ lib/enums.js
@@ -1,0 +1,14 @@
+export const tableHeaderNames = {
+  CAMPAIGN: 'campaign',
+  COUNTRY_USER: 'country-user',
+  ALL_USERS: 'all-users',
+  ALL_COUNTRIES: 'all-countries',
+  USER: 'user',
+  TEAM: 'team',
+  ADMIN_USER: 'admin-user',
+  ADMIN_CAMPAIGN_TEAMS: 'admin-campaign-teams',
+  ADMIN_CAMPAIGN_FAVORITES: 'admin-campaign-favorites',
+  ADMIN_CAMPAIGN_CONTRIBUTIONS: 'admin-campaign-contributions',
+  ADMIN_CAMPAIGN_ALL: 'admin-campaign-all',
+  TASKING_MANAGER: 'tasking-manager'
+}
